Pre-process the sub-sets
The files in this commit preprocess the sub-sets mentioned in the previous commit. Most of the code for each of these was copied from the Log contained in the Rattle GUI, but these scripts can be run without Rattle. 

I have also included the versions of the data that were fully preprocssed under /data/processed.